backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.16
langchain-openai>=0.0.5
langchain-community>=0.0.32
openai>=1.12.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

